pyproject.toml
[build-system]
requires = ["setuptools>=82.0.1,<83", "wheel>=0.46.3,<0.47"]
build-backend = "setuptools.build_meta"

[project]
name = "engllm-chat"
version = "0.1.0"
description = "Standalone interactive directory chat over repository files."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "EngLLM Contributors" }
]
dependencies = [
  "markitdown>=0.1.0,<1.0",
  "openai>=1.0.0,<2.0",
  "pathspec>=1.0.4,<2.0",
  "pydantic>=2.12.5,<3.0",
  "PyYAML>=6.0.3,<7.0",
  "textual>=0.89.1,<1.0",
]

[project.optional-dependencies]
dev = [
  "black==26.3.1",
  "isort==8.0.1",
  "ruff==0.15.6",
  "mypy==1.19.1",
  "types-PyYAML==6.0.12.20250915",
  "pytest==9.0.2",
  "pytest-cov==7.0.0",
]

[project.scripts]
engllm-chat = "engllm_chat.cli.main:main"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["engllm_chat"]
skip = [".venv", ".mypy_cache", ".pytest_cache", ".ruff_cache", "build", "dist"]

[tool.ruff]
line-length = 88
target-version = "py311"
respect-gitignore = true
src = ["src"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",
  "I",
  "B",
  "UP",
]
ignore = [
  "E501",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
packages = ["engllm_chat"]
mypy_path = ["src"]
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --cov=engllm_chat --cov-report=term-missing"

[tool.coverage.run]
source = ["engllm_chat"]
branch = true
omit = [
  "src/engllm_chat/__main__.py",
]
disable_warnings = [
  "module-not-measured",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 90

[tool.setuptools.packages.find]
where = ["src"]
include = ["engllm_chat*"]
